Refactor samhsa adapters
