+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/adapters/SamhsaHHAAdapter.java
@@ -1,0 +1,45 @@
+package gov.cms.bfd.pipeline.sharedutils.adapters;
+
+import gov.cms.bfd.model.rif.entities.HHAClaim;
+import gov.cms.bfd.model.rif.entities.HHAClaimLine;
+import gov.cms.bfd.pipeline.sharedutils.model.SamhsaFields;
+import java.lang.reflect.InvocationTargetException;
+import java.util.List;
+
+/** Adapter to get a list of SAMHSA codes from a HHAClaim. */
+public class SamhsaHHAAdapter extends SamhsaAdapterBase<HHAClaim, HHAClaimLine> {
+  /**
+   * Constructor.
+   *
+   * @param claim The claim to process.
+   * @param claimLines The claim's claim lines.
+   */
+  public SamhsaHHAAdapter(HHAClaim claim, List<HHAClaimLine> claimLines) {
+    super(claim, claimLines);
+    this.claim = claim;
+    this.claimLines = claimLines;
+    this.table = "hha_claims";
+    this.linesTable = "hha_claim_lines";
+  }
+
+  /**
+   * Retrieves a list of SAMHSA fields.
+   *
+   * @return {@link SamhsaFields}
+   * @throws InvocationTargetException Thrown when the one of the claim's methods cannot be invoked.
+   * @throws NoSuchMethodException Thrown when one of the claim's method does not exist.
+   * @throws IllegalAccessException Thrown on illegal access invoking the claim's method.
+   */
+  @Override
+  public List<SamhsaFields> getFields()
+      throws InvocationTargetException, NoSuchMethodException, IllegalAccessException {
+    getIcdDiagnosisCodes(25);
+    getPrincipalDiagnosis();
+    getDiagnosisExternalCodes(12);
+    getDiagnosisFirstCode();
+    getApcOrHippsCode();
+    getHcpcsCode();
+
+    return samhsaFields;
+  }
+}
