--- apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/s3/adapters/SamhsaCarrierAdapter.java
@@ -1,26 +1,0 @@
-package gov.cms.bfd.pipeline.sharedutils.s3.adapters;
-
-import gov.cms.bfd.model.rif.entities.CarrierClaim;
-import gov.cms.bfd.model.rif.entities.CarrierClaimLine;
-import gov.cms.bfd.pipeline.sharedutils.model.SamhsaFields;
-
-import java.lang.reflect.InvocationTargetException;
-import java.util.List;
-
-public class SamhsaCarrierAdapter extends SamhsaAdapterBase<CarrierClaim, CarrierClaimLine> {
-    public SamhsaCarrierAdapter(CarrierClaim claim, List<CarrierClaimLine> claimLines) {
-        super(claim, claimLines);
-      this.claim = claim;
-      this.claimLines = claimLines;
-      this.table = "carrier_claims";
-      this.linesTable = "carrier_claim_lines";
-   }
-   @Override
-    public List<SamhsaFields> getFields() throws InvocationTargetException, NoSuchMethodException, IllegalAccessException {
-       getIcdDiagnosisCodes(12);
-       getPrincipalDiagnosis();
-        getDiagnosisCode();
-        getHcpcsCode();
-       return samhsaFields;
-   }
-}
