--- apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/s3/adapters/SamhsaInpatientAdapter.java
@@ -1,32 +1,0 @@
-package gov.cms.bfd.pipeline.sharedutils.s3.adapters;
-
-import gov.cms.bfd.model.rif.entities.DMEClaim;
-import gov.cms.bfd.model.rif.entities.DMEClaimLine;
-import gov.cms.bfd.model.rif.entities.InpatientClaim;
-import gov.cms.bfd.model.rif.entities.InpatientClaimLine;
-import gov.cms.bfd.pipeline.sharedutils.model.SamhsaFields;
-
-import java.lang.reflect.InvocationTargetException;
-import java.util.List;
-
-public class SamhsaInpatientAdapter extends SamhsaAdapterBase<InpatientClaim, InpatientClaimLine> {
-    public SamhsaInpatientAdapter(InpatientClaim claim, List<InpatientClaimLine> claimLines) {
-        super(claim, claimLines);
-        this.claim = claim;
-      this.claimLines = claimLines;
-      this.table = "inpatient_claims";
-      this.linesTable = "inpatient_claim_lines";
-   }
-   @Override
-    public List<SamhsaFields> getFields() throws InvocationTargetException, NoSuchMethodException, IllegalAccessException {
-       getIcdDiagnosisCodes(25);
-       getDiagnosisExternalCodes(12);
-       getProcedureCodes(25);
-       getClaimDRGCode();
-       getPrincipalDiagnosis();
-       getDiagnosisAdmittingCode();
-       getHcpcsCode();
-       return samhsaFields;
-   }
-
-}
